Refactor renderer frame processing
