+++ docs/research/render_pipeline_refactor.md
@@ -1,19 +1,31 @@
-# Render pipeline refactor note
+# Render pipeline frame processor refactor
 
 ## Summary
 
-The render-loop composition and surface-merging logic now lives in a dedicated
-`RenderPipeline` helper so `GameLoop` can focus on initialization, event handling,
-and device updates. The pipeline retains the renderer scheduling, surface caching,
-and render-variant selection behaviors previously embedded in the loop.
+Refactor per-renderer frame preparation, execution, and logging into a dedicated
+`RendererFrameProcessor`. The goal is to keep `RenderPipeline` focused on
+render dispatch, executor management, and surface composition while preserving
+existing timing and logging behavior.
 
-## Sources
+## Motivation
 
-- `src/heart/runtime/render_pipeline.py` (new render pipeline helper)
-- `src/heart/runtime/game_loop.py` (game loop orchestration)
-- `src/heart/loop.py` (renderer variant selection)
-- `src/heart/environment.py` (legacy import shim updates)
+`RenderPipeline` previously mixed composition orchestration with low-level
+renderer frame handling and logging. Splitting the frame processing logic into a
+separate module reduces the number of concerns inside the pipeline and makes it
+simpler to evolve per-renderer behavior without touching dispatch logic.
+
+## Implementation notes
+
+- Added `RendererFrameProcessor` to encapsulate surface preparation, renderer
+  initialization, per-frame execution, timing updates, and per-renderer logging.
+- Updated `RenderPipeline` to delegate per-renderer work and queue depth updates
+  to the new processor while keeping merge strategy selection and composition
+  behavior intact.
+- Preserved timing snapshots by sharing the same `RendererTimingTracker`
+  instance between the processor and the pipeline.
 
 ## Materials
 
-- None
+- `src/heart/runtime/render_pipeline.py`
+- `src/heart/runtime/rendering/renderer_processor.py`
+- `src/heart/runtime/rendering/timing.py`
